Split cloud-init component to diffrent logic and view components
