+++ frontend/packages/kubevirt-plugin/locales/en/kubevirt-plugin.json
@@ -171,23 +171,23 @@
   "cloud init help": "cloud init help",
   "You can use cloud-init for post installation configuration of the guest operating system. The guest OS needs to have the cloud-init service running.": "You can use cloud-init for post installation configuration of the guest operating system. The guest OS needs to have the cloud-init service running.",
   "cloud-init is already configured in cloud images of Fedora and RHEL": "cloud-init is already configured in cloud images of Fedora and RHEL",
-  "Authorized keys must folow the SSH Pulic key format,<1></1>": "Authorized keys must folow the SSH Pulic key format,<1></1>",
-  "Invalid SSH public key format (use rfc4253 ssh-rsa format).": "Invalid SSH public key format (use rfc4253 ssh-rsa format).",
-  "Hostname": "Hostname",
-  "Authorized SSH Key": "Authorized SSH Key",
-  "SSH Key {{uiIDX}}": "SSH Key {{uiIDX}}",
-  "Add SSH Key": "Add SSH Key",
-  "Data loss confirmation": "Data loss confirmation",
-  "When using the Cloud-init form, custom values can not be applied and will be discarded.": "When using the Cloud-init form, custom values can not be applied and will be discarded.",
-  "The following fields will be kept: {{ keys }}.": "The following fields will be kept: {{ keys }}.",
-  "Are you sure you want to want to take this action?": "Are you sure you want to want to take this action?",
-  "Confirm": "Confirm",
   "Cloud-init volume exists but is not editable.": "Cloud-init volume exists but is not editable.",
   "Partial data available prior to the import": "Partial data available prior to the import",
   "This wizard shows a partial data set. A complete data set is available for viewing when you complete the import process.": "This wizard shows a partial data set. A complete data set is available for viewing when you complete the import process.",
   "Form": "Form",
   "Custom script": "Custom script",
   "Base-64 encoded": "Base-64 encoded",
+  "Authorized keys must folow the SSH Pulic key format,<1></1>": "Authorized keys must folow the SSH Pulic key format,<1></1>",
+  "Invalid SSH public key format (use rfc4253 ssh-rsa format).": "Invalid SSH public key format (use rfc4253 ssh-rsa format).",
+  "Hostname": "Hostname",
+  "Authorized SSH Key": "Authorized SSH Key",
+  "SSH Key {{uiIDX}}": "SSH Key {{uiIDX}}",
+  "Add SSH Key": "Add SSH Key",
+  "When using the Cloud-init form, custom values can not be applied and will be discarded.": "When using the Cloud-init form, custom values can not be applied and will be discarded.",
+  "The following fields will be kept: {{ keys }}.": "The following fields will be kept: {{ keys }}.",
+  "Are you sure you want to want to take this action?": "Are you sure you want to want to take this action?",
+  "Data loss confirmation": "Data loss confirmation",
+  "Confirm": "Confirm",
   "Maximum allowed size is 10 MiB": "Maximum allowed size is 10 MiB",
   "Read failed": "Read failed",
   "Obtaining the CA certificate": "Obtaining the CA certificate",
